Implemented client filtering to make NSLogger easier to use in teams
Add a small bit of info in the published Bonjour TXT record to indicating that the Bonjour service name was customized on the desktop. Clients can use the LoggerSetupBonjour() API to restrict connection to a specific desktop viewer. This way, multiple viewers can run at the same time on the network, and each developer can have its app direct logs to a specific viewer.