Implement peer-to-peer bonjour browsing on the client
Unfortunately, Apple forgot to update the CFNetServices API with peer-to-peer browsing when they added NSNetServiceBrowser's includesPeerToPeer property in iOS 7.0 / OS X 10.10. It is possible to trigger it using the private `_CFNetServiceBrowserSearchForServices` function though.

The only way to get peer-to-peer browsing without undocumented API is with the NSNetServiceBrowser class. This would require a big rewrite of LoggerClient.m and make `ALLOW_COCOA_USE` pretty much mandatory.