Do not convert C strings to UTF-8
Today, the `__FILE__` predefined macro is already encoded in UTF-8. Maybe that was not the case in 2010 when the file/line/function support was introduced in commit d43596b5753bd3d96886fd6d9cc61283cbd9edae.

Here are screenshots of what happens before and after this commit using the following code:

```c
LogMessageRawToF(logger, __FILE__, __LINE__, __PRETTY_FUNCTION__, NULL, 2, @"👋 Hello, world!");
```

Before this commit, the file is not displayed because of the double UTF-8 encoding:

<img width="1442" alt="Screenshot before fixing the encoding issue" src="https://github.com/fpillet/NSLogger/assets/51363/d84249cc-4df8-4a84-9c54-5cef42b10b3b">

After this commit, the file is properly displayed:

<img width="1442" alt="Screenshot after fixing the encoding issue" src="https://github.com/fpillet/NSLogger/assets/51363/ce35cd4c-ea01-412e-8666-8e63b5c27796">